+++ src/lib/libast/sfio/sfflsbuf.c
@@ -39,7 +39,7 @@ ptrdiff_t _sfflsbuf(Sfio_t*	f,	/* write out the buffered content of this stream 
 
 	GETLOCAL(f,local);
 
-	for(written = 0;; f->mode &= ~SFIO_LOCK)
+	for(written = 0;; f->mode &= (uint32_t)~SFIO_LOCK)
 	{	/* check stream mode */
 		if(SFMODE(f,local) != SFIO_WRITE && _sfmode(f,SFIO_WRITE,local) < 0)
 			return -1;
